Refactor all to typescript
