+++ .gitignore
@@ -2,4 +2,4 @@
 bundle.js*
 bundle.min.js
 node_modules
-dist/
+dist
