+++ src/helpers.ts
@@ -1,0 +1,74 @@
+export const emojis = [
+    "😂",
+    "😍",
+    "🤩",
+    "🤣",
+    "😊",
+    "😭",
+    "😘",
+    "😅",
+    "😁",
+    "😢",
+    "🤔",
+    "😆",
+    "🙄",
+    "😉",
+    "☺️",
+    "🤗",
+    "😔",
+    "😎",
+    "😇",
+    "🤭",
+    "😱",
+    "😌",
+    "👍",
+    "👏",
+    "👋",
+    "🙌",
+    "✌️",
+    "👌",
+    "🙏",
+    "🔥",
+    "🎉",
+    "💯",
+    "⚡️",
+    "❤️",
+    "💔",
+    "💖",
+    "💙",
+    "🌹",
+    "🌸",
+    "🎶",
+    "🤦",
+    "🤷",
+    "✨",
+    "💪",
+    "😋",
+    "💗",
+    "💚",
+    "😏",
+    "💛",
+    "🙂",
+    "💓",
+    "😄",
+    "😀",
+    "🖤",
+    "😃",
+    "🙈",
+    "👇",
+    "😒",
+    "❣️",
+  ];
+  
+  export const animations = [
+    { name: "Idle", animation: 0 },
+    { name: "Dance", animation: 3 },
+    { name: "Wave", animation: 1 },
+    { name: "Sitting", animation: 5 },
+    { name: "Spin", animation: 6 },
+    { name: "Savage", animation: 7 },
+    { name: "Kick", animation: 8 },
+    { name: "Uprock", animation: 9 },
+    { name: "Floss", animation: 10 },
+    { name: "Backflip", animation: 11 },
+  ];
