+++ tsconfig.json
@@ -1,0 +1,23 @@
+{
+    "compilerOptions": {
+      "lib": ["es2020", "dom"],
+      "module": "commonjs",
+      "target": "es2019",
+  
+      "esModuleInterop": true,
+      "strict": true,
+      "moduleResolution": "node",
+      "sourceMap": true,
+      "outDir": "dist",
+      "baseUrl": "./src/",
+      "skipLibCheck": true,
+  
+      "declaration": true,
+      "allowSyntheticDefaultImports": true,
+      "resolveJsonModule": true,
+
+    },
+    "include": ["src/**/*"],
+    "exclude": ["node_modules"]
+  }
+
